+++ CHANGELOG.md
@@ -33,6 +33,7 @@
  * improved single end support and sensitivity to respective connections (#38, #139)
  * forward None values for options to step declaration (#140)
  * raise exeption if an unknown configuration key is used
+ * configure cluster default options (#76)
 
 **additional stuff**
  * updated documentation and resolved sphinx warnings